requirements/base.txt
Django==4.2.5
django-environ==0.11.2
psycopg2==2.9.9
django-tailwind==3.6.0
crispy-tailwind==0.5.0
django-crispy-forms==2.0
django-browser-reload==1.11.0
django-compressor==4.4
django-debug-toolbar==4.2.0
sentry-sdk==1.31.0
django-storages[google]==1.14.1
django-dbbackup==4.0.2
django-anymail==10.1
django-filter==23.3
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
